Listview item onclick
Added an onClick method to log the article number when it's clicked on
within the listview.
This is to test dynamically selecting an article so a more detailed view
can be opened in a new fragment/activity. Also set the default
orientation to portrait so it cannot be changed when the device is
rotated.