Fixed tabbed layout fragments
Fixed an error causing the fragments used in the tabbed layout to be
dropped and recreated.
I changed the setoffscreenpagelimit to 2 instead of the default 1. This
allows all three tabbed fragments to be created on start and remain in
memory, allowing the same instance to be used. Previously tab 3 would be
dropped when navigating to tab 1, now this issue is resolved.