ArrayAdapter item click loads fragment
Added an onclick method for the items within the articles array to allow
a fragment to be created when clicked. This fragment receives the
article number, sent via the bundle, and is then logged in fragment 4.
This hides the tabbed layout menu and overlays the initial tabbed
fragments; it's going to be used to show details on the articles
selected by the user.
A close button on fragment 4 has also been created to close this and
revert back to viewing the tabbed fragments.