Most viewed tab
Updated the most viewed tab to get its data from a php script via an
asynctask. This script gets a count on how many times each article has
been viewed and stores the top 5 article numbers in an array. This array
is then used to get the articles from the hosted json file and return
them to the app.