Loading data into location tab.
Set up the location tab to load data into a listview using the
asynctask. This does so based on a location entered to the asynctask.The
location then loads data from a separate json file. This will soon be
updated to get the devices location and use that to get the correct
data.