Get last known location
A location listener has been set up to constantly receive updates for
the devices location., updating the last known location. This is then
ran through a method that uses a geocoder to extract the city from the
location. The location is then displayed in the log when a test button
is pressed. This will help when refreshing the data in the listview.