Rearranged methods
Updated when certain code elements are called. The onCreate method
requests permission to access location, then if it's granted it requests
location updates. The onStart method then rusn the loadData method,
getting the last known location and logging it, the same as refreshing
does. This onStart method runs everytime the app is started, eg. going
back to home, then reopening it.
A potential issue which needs testing on a physical device is that the
onCreate method may not load the initial location before the onStart
method is ran. Meaning the onStart method may load the last known
location as null or as an outdated value. This needs to wait until the
onCreate method has got a lock on the current location.