Loads fragment1 data based on location
Fragment1 now passes the location to the GetArticles asynctask. This is
then used to load the correct json data file by adding the location
string to the url. This then allows the correc file to be downloaded and
added to the correct listview. The correct listview is decided by
checking which fragment called the asynctask, using "is instance of".
This is needed as both fragment1 and fragment3 use the same asynctask
and each have a separate listview.