Location Permission Error catching
Updated the way permissions are requested and checked. The app now
checks onStart if location permissions are granted, if they are,
location updates are requested and data is loaded. If not the app
requests them. The onRequestPermissionsResults method has been
implemented to check if the user granted the permisions when asked, if
so data is loaded.
When swipe refreshing the data, the user is prompted to grant
permissions if they're not already. The same process as the onStart
implements.