SQLite DB & Handler thread
Altered the data flow of the app. Currenlty the app downloads data
within an async task called from the main activity and stores it in a
SQLite database. The correct data will then be added to the
corresponding listview in each fragment.
The location listener has been moved to a handler thread as this solved
the problem of frame skipping. This occurred when data was being
downloaded alongside the location listener on an extremely slow internet
connection resulting in a timeout.
Additionally, checks for airplane mode and enabled location services
have been implemented.