Displaying new & historic data + listview emptyview
Implemented the use of downloaded data, displaying in the correct
listview once downloaded. If there's errors downloading data, an error
message is displayed and historic data from the SQLite database is added
to the required listviews.
If there's no data available (either none to download or none in
database), a textview emptyview is displayed in the listview stating no
data available.