Load all data + errors
The app now attempts to download all data onStart. If it can't data is
added to the listviews from the SQLite database.

When data has been downloaded, all and top data are added instantly, but
a check for location data has been added to see if it returned null. If
null was returned, data is removed from the database and and the list
empty view is changed, stating 'Location data is not available at this
time. There may be no articles in your location.'. The list is then set
to use data from the database which is ofcourse empty, resulting in the
the updated empty view message being displayed.
This enables all and top data to be downloaded and displayed regardless
of whether the location data returns no data.