Tabs swipe refresh & fixing bugs
Swiping down on refresh layouts for each tab now redownloads the
corresponding data. The app now checks if the location url can be
accessed before attempting to download. If it doesn't exist then the
list is set to null and location string is passed through to
onPostExecute. This then checks if the location string is null or just a
city that no data is stored for. Correct empty view textviews are
displayed (either cannot acquire location, or no data for current
location).