Location shared prefs
Stores location in shared preferences everytime the device moves to a
new town/city. Implemented by getting continuous location updates and
using a geocoder to get the city, comparing it to the previously stored
city. This keeps the last known location up to date constantly.
This value is then used to load location based articles via the
asynctask.